Split up admin.go code to organize it better
